pyproject.toml
[tool.poetry]
name = "1"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
pillow = "^10.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
redis = "^5.0.1"
routers = "^0.10.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
generics = "^6.0.0"
coverage = "^7.4.1"
coreapi = "^2.3.3"
pyyaml = "^6.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
